requirements.txt
pillow
opencv-python
pillow-heif
numpy
